feat: update type

+++ main.cpp
@@ -5,36 +5,36 @@
 #include <iostream>
 #include <vector>
 
-long solve(long n, std::vector<std::array<long, 2>> arr) {
+long long solve(long long n, std::vector<std::array<long long, 2>> arr) {
    return 0;
 }
 
 
 int main() {
-    long k;
+    long long k;
     std::cin >> k;
 
-    for (long i = 0; i < k; i++) {
-        long n;
+    for (long long i = 0; i < k; i++) {
+        long long n;
         std::cin >> n;
 
-        std::vector<std::array<long, 2>> arr;
-        for (long i = 0; i < n; i++) {
-            long x;
+        std::vector<std::array<long long, 2>> arr;
+        for (long long i = 0; i < n; i++) {
+            long long x;
             std::cin >> x;
-            std::array<long, 2> elm;
+            std::array<long long, 2> elm;
             elm.fill(x);
 
             arr.push_back(elm);
         }
 
-        for (long i = 0; i < n; i++) {
-            long y;
+        for (long long i = 0; i < n; i++) {
+            long long y;
             std::cin >> y;
             arr[i][1] = y;
         }
 
-        long ans = solve(n, arr);
+        long long ans = solve(n, arr);
         std::cout << ans << std::endl;
     }
     return 0;
